test(ts): unify `Remove User` benchmark
